+++ msaBase/configurate.py
@@ -6,6 +6,7 @@ Initialize with a MSAServiceDefintion Instance to control the features and funct
 """
 import json
 import os
+import aiohttp
 from asyncio import Task
 from datetime import datetime
 from functools import wraps
@@ -178,6 +179,42 @@ class MSAApp(FastAPI):
         self.add_event_handler("startup", self.startup_event)
         self.create_dapr_endpoint()
 
+    def save_config(self, config: ConfigInput) -> None:
+        """
+        Save config to file.
+
+        Parameters:
+
+            config: config for save.
+        """
+
+        if config.data.config.name == self.settings.name:
+            reload_needed = self.update_settings(config.data.config, config.data.one_time)
+            if reload_needed:
+                self.logger.info("New config needs reload.")
+                with open("config.json", "w") as json_file:
+                    json.dump(config.data.dict(), json_file)
+
+                self.logger.info("New config saved to config.json")
+
+    async def load_config(self, url: str) -> None:
+        """
+        Get config.
+
+        Parameters:
+
+            url: request URL.
+        """
+
+        async with aiohttp.ClientSession() as session:
+            async with session.get(url) as resp:
+                if resp.status == 200:
+                    config = MSAServiceDefinition.parse_obj(await resp.json())
+                    data = ConfigDTO(config=config, one_time=False)
+                    self.save_config(config=ConfigInput(data=data))
+                else:
+                    self.logger.info(f"Config not found")
+
     def create_dapr_endpoint(self):
         """
         Subscribes service to pubsub topic through which new configs will be received.
@@ -193,14 +230,7 @@ class MSAApp(FastAPI):
             """
             try:
                 self.logger.info(f"Received config from spkRegistry. Data: {received_config.data}")
-                if received_config.data.config.name == self.settings.name:
-                    reload_needed = self.update_settings(received_config.data.config, received_config.data.one_time)
-                    if reload_needed:
-                        self.logger.info("New config needs reload.")
-                        with open("config.json", "w") as json_file:
-                            json.dump(received_config.data.dict(), json_file)
-
-                        self.logger.info("New config saved to config.json")
+                self.save_config(config=received_config)
 
             except Exception as ex:
                 self.logger.info(ex)